fastapi_server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
backoff==2.2.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fonttools==4.58.1
frozenlist==1.6.2
google-api-core==2.25.0
google-auth==2.40.2
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.72.1
h11==0.16.0
idna==3.10
inference-sdk==0.50.3
kiwisolver==1.4.8
marshmallow==3.26.1
matplotlib==3.10.3
multidict==6.4.4
mypy_extensions==1.1.0
numpy==2.2.6
opencv-python==4.10.0.84
packaging==25.0
passlib==1.7.4
pillow==11.2.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
supervision==0.25.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
websockets==15.0.1
yarl==1.20.0
